spotify implemented in the result page and working
